Update Board.java
Fixed the checkmate detection so that it is working properly. Added a method to grab a piece of a specific color given a coordinate in order to eliminate the problem of the makeMove method moving the wrong piece when two pieces are on the same square in the generateMoves method.